feat(core): Projekt-Lader mit Migration
